backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "google-adk>=1.7.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2.5.1",
    "langchain>=0.1.0",
    "langchain-community>=0.0.18",
    "chromadb>=1.0.15",
    "pypdf>=5.8.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.6.2",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "llama-index-vector-stores-qdrant>=0.1.0",
    "google-generativeai>=0.3.0",
    "pypdf2>=3.0.1",
    "python-magic>=0.4.27",
    "aiofiles>=23.2.1",
    "nest-asyncio>=1.5.8",
    "python-dateutil>=2.8.20",
    "elevenlabs>=0.2.26",
    "gtts>=2.4.0",
    "faster-whisper>=0.10.0",
    "email-validator>=2.2.0",
]
